Move MemorySegment to its own 'memory' module
